refactor: extract 4 submodules, deduplicate challenge macro, decompose large render functions
- Extract combat/enemy_generation.rs from types.rs (~430 LOC)
- Extract morris/ai.rs from morris/logic.rs (~370 LOC)
- Extract ui/enemy_sprite_data.rs from enemy_sprites.rs (~640 LOC)
- Create impl_apply_game_result! macro deduplicating 10 challenge files (~300 LOC saved)
- Extract apply_xp_and_check_levelup() helper in tick_stages.rs (8x duplication removed)
- Decompose render_play_field() in flappy_scene.rs (500 LOC -> 5 sub-functions)
- Decompose draw_water_scene() in fishing_scene.rs (406 LOC -> 4 sub-functions)
