+++ src/combat/mod.rs
@@ -3,6 +3,7 @@
 pub mod attacks;
 pub(crate) mod damage;
 pub(crate) mod enemy_attack;
+pub mod enemy_generation;
 pub mod events;
 pub mod logic;
 pub mod orchestration;
