+++ src/ui/enemy_sprites.rs
@@ -1,8 +1,16 @@
-/// Enemy sprite templates for 3D rendering
+//! Enemy sprite rendering: struct definitions, archetype enum, color palettes,
+//! and sprite selection logic.
+//!
+//! Sprite constant data (legacy ASCII, pixel art, suffix mapping tables)
+//! lives in `enemy_sprite_data`.
+
 use ratatui::style::Color;
 
 use crate::core::game_state::GameState;
 use crate::zones::get_zone;
+
+// Re-export all sprite data so existing internal consumers still compile.
+pub use super::enemy_sprite_data::*;
 
 #[allow(dead_code)]
 pub struct EnemySprite {
@@ -22,428 +30,9 @@ impl EnemySprite {
     }
 }
 
-// ── 8 Base Sprite Archetypes (legacy ASCII art, kept for reference) ──────────
-
-#[allow(dead_code)]
-pub const SPRITE_INSECT: EnemySprite = EnemySprite::new(
-    r"    ╲│╱  ╲│╱
-      ╲╱  ╲╱
-     ┌──────┐
-    ╱│ ●  ● │╲
-   ╱ │  ▼▼  │ ╲
-   ╲ │▒▒▒▒▒▒│ ╱
-    ╲└──────┘╱
-    ╱├──────┤╲
-   ╱ ╰──────╯ ╲
-  ╱╱            ╲╲",
-    16,
-    10,
-);
-
-#[allow(dead_code)]
-pub const SPRITE_QUADRUPED: EnemySprite = EnemySprite::new(
-    r"   ╱▲    ▲╲
-  ╱  ╱╲  ╱╲  ╲
- │  ● ╱██╲ ●  │
- │   ╱████╲   │
- │  │ ▼══▼ │  │
-  ╲ ╰══════╯ ╱
-   ╲ ██████ ╱
-    ▐██████▌
-   ╱╱ ╱╲╱╲ ╲╲
-  ╰╯ ╰╯  ╰╯ ╰╯",
-    16,
-    10,
-);
-
-#[allow(dead_code)]
-pub const SPRITE_SERPENT: EnemySprite = EnemySprite::new(
-    r"       ╱╲
-    ╱▓▓▓▓╲
-   │ ◆  ◆ │
-   │  ╲╱╲  │
-   ╰┐ ▼▼ ┌╯
-  ╱▓╰════╯▓╲
- │▓▓╲      ╱▓│
-  ╲▓▓╲  ╱╱▓╱
-   ╲▓▓╲╱╱▓╱
-    ╰══╲╱══╯",
-    15,
-    10,
-);
-
-#[allow(dead_code)]
-pub const SPRITE_HUMANOID: EnemySprite = EnemySprite::new(
-    r"     ╱══╲
-    ╱ ▓▓ ╲
-    │ ●  ● │
-    │  ▼   │
-    ╰┬────┬╯
-   ╱─┤████├─╲
-  ╱  │████│  ╲
-     │████│
-     ├─┬┬─┤
-     ╰─╯╰─╯",
-    15,
-    10,
-);
-
-#[allow(dead_code)]
-pub const SPRITE_AVIAN: EnemySprite = EnemySprite::new(
-    r"       ╱╲
-╲     ╱████╲     ╱
- ╲   │ ◆  ◆ │   ╱
-  ╲  │  ╲╱   │  ╱
-   ╲ ╰──────╯ ╱
-    ╲ ▒████▒ ╱
-     ╲ ████ ╱
-      ╲▒▒▒▒╱
-      ╱╲  ╱╲
-     ╱╱ ╲╱ ╲╲",
-    17,
-    10,
-);
-
-#[allow(dead_code)]
-pub const SPRITE_ELEMENTAL: EnemySprite = EnemySprite::new(
-    r"    ╱░░░░╲
-   ╱░▒▒▒▒░╲
-  │░▒ ◆◆ ▒░│
-  │░▒▓▓▓▓▒░│
-  │░▒▓██▓▒░│
-  │░▒▓▓▓▓▒░│
-  │░▒▒▒▒▒▒░│
-   ╲░░░░░░╱
-    ╲░░░░╱
-      ░░",
-    14,
-    10,
-);
-
-#[allow(dead_code)]
-pub const SPRITE_TITAN: EnemySprite = EnemySprite::new(
-    r"   ═══════════
-   ║ ●     ● ║
-   ║    ▼▼    ║
-   ║ ╱════╲  ║
-  ╔╩════════╩╗
-  ║██████████║
-  ║██████████║
-  ╚╦════════╦╝
-   ║║      ║║
-   ╩╩      ╩╩",
-    15,
-    10,
-);
-
-#[allow(dead_code)]
-pub const SPRITE_HORROR: EnemySprite = EnemySprite::new(
-    r"   ╱╲  ╱╲  ╱╲
-  ╱ ▒▓▒▓▒▓▒ ╲
- │ ●  ◆  ● ◆│
- │ ▓▒░▒▓░▒▓ │
- │ ╱▓█▓▓█▓╲ │
-  ╲▓█▓▒▒▓█▓╱
-  ╱▒╲▓▓▓▓╱▒╲
- ╱▒╱╲╲▒▒╱╱▒╲╲
- ╲╱  ╲╲╱╱  ╲╱
-      ╲╱╲╱",
-    16,
-    10,
-);
-
-// ── Boss Crown Patterns ─────────────────────────────────────────────
-
-pub const BOSS_CROWN: &str = "--- \u{2605} ---";
-pub const ZONE_BOSS_CROWN: &str = "=== \u{2605} ===";
-
-// ── Pixel Sprite Definitions ─────────────────────────────────────────
-//
-// Pixel format per character:
-//   '.' = transparent (preserve zone background)
-//   'D' = dark outline
-//   'B' = body primary color
-//   'S' = shade (darker body)
-//   'E' = eye / accent color
-//   'H' = highlight (bright accent)
-//
-// Rows are processed in pairs (row 0+1, row 2+3, ...) to produce one
-// terminal cell row each, using ▀ / ▄ / █ half-block characters.
-
 pub struct PixelSprite {
     pub rows: &'static [&'static str],
 }
-
-// HUMANOID: Armored warrior. Crested helm, broad pauldrons, weapon-arm raised, two legs.
-// 16 pixels wide, 18 pixels tall → 9 cell rows
-pub const PIXEL_HUMANOID: PixelSprite = PixelSprite {
-    rows: &[
-        "....HDDDDH......",
-        "...DBBBBBBD.....",
-        "...DBEEBBBD.....",
-        "...DBBBBSBD.....",
-        "....DDDDDD......",
-        "..HDBBBBBDDH....",
-        ".DBBBBBBBBBBBD..",
-        "HDBSBBBBBBBBSDH.",
-        ".DBBBBBBBBBBD...",
-        ".DBSBBBBBBSBD...",
-        "..DBBHBBHBBD....",
-        "..DBBBBBBBD.....",
-        "..DBBD.DBBD.....",
-        "..DBBD.DBBD.....",
-        "..DBSD.DBSD.....",
-        "..DBBD.DBBD.....",
-        "..DSSD.DSSD.....",
-        "..DDDD.DDDD.....",
-    ],
-};
-
-// INSECT: Bug with antennae, segmented oval body, 3 pairs of jointed legs, compound eyes.
-// 18 pixels wide, 16 pixels tall → 8 cell rows
-pub const PIXEL_INSECT: PixelSprite = PixelSprite {
-    rows: &[
-        "D.....D....D.....D",
-        "..D..D......D..D..",
-        "....DBBBBBBBBD....",
-        "....DEEBBBEEBD....",
-        "....DBBBBBBBD.....",
-        ".DBBBBBBBBBBBBBD..",
-        "DBBHBBBBBBBBBBHBD.",
-        "DBSSBBBBBBBBSSBD..",
-        "DBBBBBBBBBBBBBBBD.",
-        ".DBD.DBBBBBBD.DBD.",
-        ".BD..DBBBBBD...DB.",
-        ".D...DBBBBD.....D.",
-        "....DBBBBBBD......",
-        "....DBBSSBD.......",
-        "....DBBBBD........",
-        "....DDDDD.........",
-    ],
-};
-
-// QUADRUPED: Four-legged beast. Profile view, prominent head, muscular barrel body, four legs.
-// 18 pixels wide, 18 pixels tall → 9 cell rows
-pub const PIXEL_QUADRUPED: PixelSprite = PixelSprite {
-    rows: &[
-        ".DBD.......DBD....",
-        ".DBBD.....DBSD....",
-        "..DBBBBBBBBBD.....",
-        "..DBEBBBBEBBD.....",
-        "..DBBBBBBBBBD.....",
-        "..DBHBBBBHBBD.....",
-        "..DBBBBBBBBBD.....",
-        ".DBBSSBBBBSSBD....",
-        ".DBBBBBBBBBSBD....",
-        ".DBBBBBBBBBBD.....",
-        "..DBBBBBBBBD......",
-        "..DBD...DBBD......",
-        ".DBD....DSBD......",
-        ".DBD....DBBD......",
-        ".DSD....DSSD......",
-        ".DDD....DDDD......",
-        "..................",
-        "..................",
-    ],
-};
-
-// SERPENT: Cobra strike pose, flared hood, S-curve body with scale pattern.
-// 16 pixels wide, 18 pixels tall → 9 cell rows
-pub const PIXEL_SERPENT: PixelSprite = PixelSprite {
-    rows: &[
-        "....HDDDDH......",
-        "...DBBBBBBBD....",
-        "..DBBBBBBBBBD...",
-        ".DBBEBBBBEBBD...",
-        "..DBBBBBBBBD....",
-        "...DBBEBBD......",
-        "....DBBBD.......",
-        ".....DBBD.......",
-        "..DBBBBD........",
-        ".DBSBBBD........",
-        ".DBBBBD.........",
-        "......DBBBBD....",
-        ".....DBSBBBD....",
-        "......DBBBD.....",
-        "..DBBBD.........",
-        "...DBBBD........",
-        "....DBSD........",
-        ".....HDD........",
-    ],
-};
-
-// AVIAN: Front-facing raptor. Wide M-wingspan, hooked beak, layered feathers, spread talons.
-// 20 pixels wide, 18 pixels tall → 9 cell rows
-pub const PIXEL_AVIAN: PixelSprite = PixelSprite {
-    rows: &[
-        "........DDDDD.......",
-        ".......DBBBBBD......",
-        ".......DBEBBBD......",
-        ".......DBHBBD.......",
-        "....DDDDBBBBDDDD....",
-        "..DHBBBBBBBBBBBHBD..",
-        ".DBBSSBBBBBBBBSSBBHD",
-        "DBBBBBBBBBBBBBBBBBD.",
-        ".DBBSSBBBBBBBBSSBD..",
-        "...DDDBBBBBBBBDDD...",
-        "......DBBBBBBD......",
-        "......DBBBSSBD......",
-        ".......DBBBD........",
-        "......DBBD.DBD......",
-        ".....DBD....DBD.....",
-        "....DBHD....DBHD....",
-        "...DBBD......DBBD...",
-        "...DDD........DDD...",
-    ],
-};
-
-// ELEMENTAL: Amorphous energy entity. H core, irregular spiked outline, E accent sparks, asymmetric.
-// 18 pixels wide, 18 pixels tall → 9 cell rows
-pub const PIXEL_ELEMENTAL: PixelSprite = PixelSprite {
-    rows: &[
-        "....H......H......",
-        "...DHHBD...H......",
-        "..DBHHHBBD........",
-        ".DBHHHHHBBD.H.....",
-        ".DBHHEEHBBBD......",
-        "DBHHHEEBBBBD......",
-        "DBBBHHHBBBBD......",
-        ".DBBHBBBSBD.......",
-        ".DBBBBBHSBD.......",
-        "..DBBBSBBD........",
-        "..DBBBBBBD.H......",
-        "...DBHBBD.........",
-        "...DBBSBD.........",
-        "....DBBBD.........",
-        "....DBHD..........",
-        "...DBD.H..........",
-        "..DD..............",
-        "..................",
-    ],
-};
-
-// TITAN: Enormous armored colossus. Tiny head, massive pauldrons, thick barrel torso, column legs.
-// 20 pixels wide, 18 pixels tall → 9 cell rows
-pub const PIXEL_TITAN: PixelSprite = PixelSprite {
-    rows: &[
-        "......HDDDDH........",
-        "......DBBBBD........",
-        "......DBEEBD........",
-        "......DBBSBD........",
-        "..HDDDDBBBBDDDDH....",
-        "..DBBHBBBBBBHBBBD...",
-        ".DBBBBBBBBBBBBBSBD..",
-        ".DBBHBBBBBBBBHBBBD..",
-        ".DBSBBHHBBBBBBBSBD..",
-        ".DBBSBBBBBBBBSBBBD..",
-        ".DBBBBBBHBBBBBSBD...",
-        "..DBBBBBBBBBBBBD....",
-        "...DBBBD..DBBBD.....",
-        "...DBSBD..DBSBD.....",
-        "...DBBBD..DBBBD.....",
-        "...DBBBD..DBBBD.....",
-        "...DSSSD..DSSSD.....",
-        "...DDDDD..DDDDD.....",
-    ],
-};
-
-// HORROR: Asymmetric eldritch abomination. Scattered E eyes, irregular tentacles, mottled body.
-// 20 pixels wide, 18 pixels tall → 9 cell rows
-pub const PIXEL_HORROR: PixelSprite = PixelSprite {
-    rows: &[
-        "..E.........E.......",
-        "..D....E............",
-        "....DDDDDDDDDD......",
-        "...DBBBSBBEBBD......",
-        "..DBHSBBBSBBBBBD....",
-        "..DBEBBBSBBHBBBD....",
-        ".DBBBHBBBBSBBBBBBD..",
-        ".DBBSBBEBBBBBSBD.DD.",
-        ".DBBBBBSBBHBBBBD....",
-        "..DBBEBBBBBBBSBD....",
-        "..DBBBSBBHBBBD......",
-        "...DDDDDDDDDD.......",
-        "....DBBBD...........",
-        "...DBBD.DD..........",
-        "...DBD..D...........",
-        "....DD..D...........",
-        "....D...............",
-        "................E...",
-    ],
-};
-
-// UNDEAD: Skeletal warrior. Hollow eyes, exposed ribs, bony arms, leg bones.
-// 16 pixels wide, 18 pixels tall → 9 cell rows
-pub const PIXEL_UNDEAD: PixelSprite = PixelSprite {
-    rows: &[
-        "....HDDDDH......",
-        "...DBBEBBD......",
-        "...DBEEBBD......",
-        "....DDDDD.......",
-        "...DBSBSBD......",
-        "..HDBBBBBBDH....",
-        "..DSBSDBSBD.....",
-        "..DBBBBBBBD.....",
-        "...DDDDDDD......",
-        "..DBBBD.DBD.....",
-        "..DBBD..DBD.....",
-        "..DBBD..DBD.....",
-        "...DDD..DDD.....",
-        "..DBD....DBD....",
-        ".DBBD....DBBD...",
-        ".DBSD....DSBD...",
-        "..DDD....DDD....",
-        "................",
-    ],
-};
-
-// PLANT: Aggressive flora. Leafy canopy, thorned vines, root legs.
-// 16 pixels wide, 18 pixels tall → 9 cell rows
-pub const PIXEL_PLANT: PixelSprite = PixelSprite {
-    rows: &[
-        "...HBHBHBHB.....",
-        "..HBBBBBBBBD....",
-        "..DBEBEBBBD.....",
-        "..DBBBBBBBD.....",
-        ".DBHBBBBBBBD....",
-        "DBBBBBBBBBBBD...",
-        "DBSBBBHBBSBBBD..",
-        "DBBBBBBBBBBBD...",
-        ".DBBHBBHBBBD....",
-        "..DBBBBBBBD.....",
-        "DBBBBBBBBBBBBBD.",
-        "DBBSBBBBSBBBBBD.",
-        ".DSBBBBBBBSBD...",
-        "..DDBBBBBBDD....",
-        "...DBBBBBBD.....",
-        "..DBSD.DBSD.....",
-        ".DBBBD.DBBBD....",
-        "DDSSD...DSSDD...",
-    ],
-};
-
-// AQUATIC: Sea creature. Streamlined torpedo body, fin ridges, deep-water coloring.
-// 18 pixels wide, 16 pixels tall → 8 cell rows
-pub const PIXEL_AQUATIC: PixelSprite = PixelSprite {
-    rows: &[
-        "..DBBBBBBBBBBBBBD.",
-        ".DEBBBBBBBBBBBBBED",
-        "DBSBBBBBBBBBBBBSBD",
-        "DEBBBBBBBBBBBBBBBD",
-        "DBSSBBBBBBBBBBBSBD",
-        ".DBBBBBBBBBBBBBBD.",
-        "DBBSBBBBBBBBSBBD..",
-        ".DBBBBBBBBBBBBD...",
-        "DBBSSBBBBBBSSBD...",
-        ".DBBBBBBBBBBBBD...",
-        "..DBBBBBBBBBBD....",
-        "DBBBBBBBBBBBBBBBD.",
-        ".DSBBBBBBBBBSBBD..",
-        "..DBBBBBBBBBBD....",
-        "...DBBBBBBBBBD....",
-        "....DHBBBBBHD.....",
-    ],
-};
 
 // ── Sprite Archetype Enum ───────────────────────────────────────────
 
@@ -655,7 +244,7 @@ fn relative_luminance(r: u8, g: u8, b: u8) -> f32 {
     0.2126 * linearize(r) + 0.7152 * linearize(g) + 0.0722 * linearize(b)
 }
 
-/// Contrast ratio between two RGB colors (always ≥ 1.0).
+/// Contrast ratio between two RGB colors (always >= 1.0).
 fn contrast_ratio_rgb(a: (u8, u8, u8), b: (u8, u8, u8)) -> f32 {
     let la = relative_luminance(a.0, a.1, a.2) + 0.05;
     let lb = relative_luminance(b.0, b.1, b.2) + 0.05;
@@ -680,9 +269,7 @@ pub fn sprite_color_palette(zone_id: u32, tier: EnemyTier) -> SpriteColorPalette
     let (pr, pg, pb) = ansi_to_rgb(primary_color);
     let (er, eg, eb) = ansi_to_rgb(zone_pal.secondary);
 
-    // dark: ~20% of body per channel. This gives a deep tinted shadow that forms the
-    // sprite outline. For very dark primaries like Blue (0,0,170), even 20% gives low
-    // contrast — this is an inherent limitation of dark ANSI hues, and is accepted.
+    // dark: ~20% of body per channel.
     let dark = Color::Rgb(
         scale_channel(pr, 0.20),
         scale_channel(pg, 0.20),
@@ -699,9 +286,7 @@ pub fn sprite_color_palette(zone_id: u32, tier: EnemyTier) -> SpriteColorPalette
         scale_channel(pb, 0.60),
     );
 
-    // eye: secondary zone color, blended toward White when contrast against body is too
-    // low (< 1.5×). This ensures E pixels (eyes / accents) are visible even when the
-    // secondary color has similar luminance to the primary (e.g. Yellow on Green).
+    // eye: secondary zone color, blended toward White when contrast is low
     let eye_raw = (er, eg, eb);
     let eye = if contrast_ratio_rgb(eye_raw, (pr, pg, pb)) < 1.5 {
         Color::Rgb(
@@ -713,9 +298,7 @@ pub fn sprite_color_palette(zone_id: u32, tier: EnemyTier) -> SpriteColorPalette
         Color::Rgb(er, eg, eb)
     };
 
-    // highlight: ~140% of body clamped to 255. When all channels are already at max
-    // (near-white body), simple scaling produces the same color as body. In that case
-    // shift toward warm white (reduced blue) so H pixels have a visible warm shimmer.
+    // highlight: ~140% of body clamped to 255
     let hl_r = scale_channel(pr, 1.40);
     let hl_g = scale_channel(pg, 1.40);
     let hl_b = scale_channel(pb, 1.40);
@@ -731,220 +314,6 @@ pub fn sprite_color_palette(zone_id: u32, tier: EnemyTier) -> SpriteColorPalette
         shade,
         eye,
         highlight,
-    }
-}
-
-// ── Zone Suffix-to-Archetype Mapping ────────────────────────────────
-
-/// Returns the sprite archetype for a zone enemy suffix.
-/// Falls back to the zone default if the suffix is unrecognized.
-pub fn archetype_for_suffix(zone_id: u32, suffix: &str) -> SpriteArchetype {
-    let s = suffix.to_lowercase();
-    let matched = match zone_id {
-        1 => match s.as_str() {
-            "beetle" | "wasp" | "mantis" => Some(SpriteArchetype::Insect),
-            "rabbit" | "boar" | "hare" | "toad" => Some(SpriteArchetype::Quadruped),
-            "serpent" => Some(SpriteArchetype::Serpent),
-            "grub" => Some(SpriteArchetype::Insect),
-            "sprout" => Some(SpriteArchetype::Plant),
-            _ => None,
-        },
-        2 => match s.as_str() {
-            "wolf" | "lynx" => Some(SpriteArchetype::Quadruped),
-            "spider" | "moth" => Some(SpriteArchetype::Insect),
-            "bat" => Some(SpriteArchetype::Avian),
-            "treant" => Some(SpriteArchetype::Titan),
-            "wisp" => Some(SpriteArchetype::Elemental),
-            "hollow" => Some(SpriteArchetype::Horror),
-            "shambler" => Some(SpriteArchetype::Plant),
-            "wretch" => Some(SpriteArchetype::Humanoid),
-            _ => None,
-        },
-        3 => match s.as_str() {
-            "goat" | "ram" => Some(SpriteArchetype::Quadruped),
-            "eagle" | "condor" => Some(SpriteArchetype::Avian),
-            "golem" | "yeti" | "troll" | "gargoyle" => Some(SpriteArchetype::Titan),
-            "harpy" | "bandit" => Some(SpriteArchetype::Humanoid),
-            _ => None,
-        },
-        4 => match s.as_str() {
-            "skeleton" | "mummy" | "revenant" | "lich" => Some(SpriteArchetype::Undead),
-            "spirit" => Some(SpriteArchetype::Elemental),
-            "gargoyle" => Some(SpriteArchetype::Titan),
-            "specter" => Some(SpriteArchetype::Horror),
-            "shade" => Some(SpriteArchetype::Horror),
-            "cultist" => Some(SpriteArchetype::Humanoid),
-            "apparition" => Some(SpriteArchetype::Elemental),
-            _ => None,
-        },
-        5 => match s.as_str() {
-            "salamander" | "cinderwyrm" => Some(SpriteArchetype::Serpent),
-            "phoenix" | "drake" => Some(SpriteArchetype::Avian),
-            "imp" | "ashborn" => Some(SpriteArchetype::Humanoid),
-            "elemental" | "magmite" => Some(SpriteArchetype::Elemental),
-            "hellhound" => Some(SpriteArchetype::Quadruped),
-            "infernal" => Some(SpriteArchetype::Horror),
-            _ => None,
-        },
-        6 => match s.as_str() {
-            "mammoth" | "glacial" => Some(SpriteArchetype::Titan),
-            "wendigo" | "wraith" => Some(SpriteArchetype::Horror),
-            "bear" | "moose" => Some(SpriteArchetype::Quadruped),
-            "wyrm" => Some(SpriteArchetype::Serpent),
-            "banshee" => Some(SpriteArchetype::Elemental),
-            "imp" => Some(SpriteArchetype::Humanoid),
-            "revenant" => Some(SpriteArchetype::Undead),
-            _ => None,
-        },
-        7 => match s.as_str() {
-            "construct" | "golem" | "sentinel" => Some(SpriteArchetype::Titan),
-            "guardian" => Some(SpriteArchetype::Humanoid),
-            "sprite" | "shard" | "prism" => Some(SpriteArchetype::Elemental),
-            "watcher" | "echo" => Some(SpriteArchetype::Horror),
-            "crawler" => Some(SpriteArchetype::Insect),
-            _ => None,
-        },
-        8 => match s.as_str() {
-            "kraken" => Some(SpriteArchetype::Horror),
-            "shark" | "ray" | "lurker" => Some(SpriteArchetype::Aquatic),
-            "naga" => Some(SpriteArchetype::Serpent),
-            "leviathan" => Some(SpriteArchetype::Titan),
-            "siren" => Some(SpriteArchetype::Humanoid),
-            "eel" | "hydra" => Some(SpriteArchetype::Serpent),
-            "drowned" => Some(SpriteArchetype::Undead),
-            _ => None,
-        },
-        9 => match s.as_str() {
-            "griffin" | "roc" | "wyvern" | "pegasus" | "stormhawk" => Some(SpriteArchetype::Avian),
-            "djinn" | "sylph" | "zephyr" => Some(SpriteArchetype::Elemental),
-            "manticore" => Some(SpriteArchetype::Quadruped),
-            "cloudwalker" => Some(SpriteArchetype::Humanoid),
-            _ => None,
-        },
-        10 => match s.as_str() {
-            "titan" | "colossus" | "juggernaut" | "breaker" => Some(SpriteArchetype::Titan),
-            "lord" | "king" | "champion" | "warlord" | "stormknight" => {
-                Some(SpriteArchetype::Humanoid)
-            }
-            "thunderborn" => Some(SpriteArchetype::Elemental),
-            _ => None,
-        },
-        11 => match s.as_str() {
-            "beast" => Some(SpriteArchetype::Quadruped),
-            "horror" | "terror" | "abomination" | "rift" | "amalgam" => {
-                Some(SpriteArchetype::Horror)
-            }
-            "fiend" => Some(SpriteArchetype::Humanoid),
-            "monster" => Some(SpriteArchetype::Titan),
-            "void" => Some(SpriteArchetype::Elemental),
-            "remnant" => Some(SpriteArchetype::Undead),
-            _ => None,
-        },
-        _ => None,
-    };
-
-    matched.unwrap_or(zone_default_archetype(zone_id))
-}
-
-/// Returns the default archetype for a zone (used when suffix doesn't match).
-fn zone_default_archetype(zone_id: u32) -> SpriteArchetype {
-    match zone_id {
-        1 | 2 | 6 => SpriteArchetype::Quadruped,
-        3 => SpriteArchetype::Titan,
-        4 | 10 => SpriteArchetype::Humanoid,
-        5 | 7 => SpriteArchetype::Elemental,
-        8 => SpriteArchetype::Serpent,
-        9 => SpriteArchetype::Avian,
-        11 => SpriteArchetype::Horror,
-        _ => SpriteArchetype::Quadruped,
-    }
-}
-
-/// Archetype matching for dungeon enemies with generic names (Orc, Troll, etc.)
-fn dungeon_generic_archetype(suffix: &str) -> SpriteArchetype {
-    match suffix.to_lowercase().as_str() {
-        "orc" => SpriteArchetype::Humanoid,
-        "troll" => SpriteArchetype::Titan,
-        "drake" => SpriteArchetype::Avian,
-        "crusher" => SpriteArchetype::Titan,
-        "beast" | "fiend" => SpriteArchetype::Quadruped,
-        "horror" | "terror" => SpriteArchetype::Horror,
-        "render" | "maw" => SpriteArchetype::Horror,
-        _ => SpriteArchetype::Quadruped,
-    }
-}
-
-/// Archetype matching for boss enemies using keyword matching on the boss name.
-fn boss_name_archetype(boss_name: &str) -> Option<SpriteArchetype> {
-    let name = boss_name.to_lowercase();
-    // Check specific creature keywords first (before generic title keywords)
-    if name.contains("spider") || name.contains("sporeling") || name.contains("arachne") {
-        Some(SpriteArchetype::Insect)
-    } else if name.contains("wolf") || name.contains("bear") || name.contains("beast") {
-        Some(SpriteArchetype::Quadruped)
-    } else if name.contains("treant")
-        || name.contains("giant")
-        || name.contains("golem")
-        || name.contains("colossus")
-        || name.contains("titan")
-        || name.contains("mammoth")
-        || name.contains("leviathan")
-        || name.contains("behemoth")
-    {
-        Some(SpriteArchetype::Titan)
-    } else if name.contains("wyrm")
-        || name.contains("serpent")
-        || name.contains("naga")
-        || name.contains("salamander")
-    {
-        Some(SpriteArchetype::Serpent)
-    } else if name.contains("revenant") || name.contains("lich") || name.contains("mummy") {
-        Some(SpriteArchetype::Undead)
-    } else if name.contains("shark") || name.contains("lurker") || name.contains("ray") {
-        Some(SpriteArchetype::Aquatic)
-    } else if name.contains("shambler") || name.contains("sprout") {
-        Some(SpriteArchetype::Plant)
-    } else if name.contains("horror")
-        || name.contains("wraith")
-        || name.contains("specter")
-        || name.contains("kraken")
-        || name.contains("avatar")
-        || name.contains("frozen one")
-        || name.contains("drowned")
-        || name.contains("broodmother")
-    {
-        Some(SpriteArchetype::Horror)
-    } else if name.contains("drake")
-        || name.contains("phoenix")
-        || name.contains("harpy")
-        || name.contains("roc")
-    {
-        Some(SpriteArchetype::Avian)
-    } else if name.contains("elemental")
-        || name.contains("wisp")
-        || name.contains("sprite")
-        || name.contains("storm")
-        || name.contains("tempest")
-        || name.contains("incarnate")
-    {
-        Some(SpriteArchetype::Elemental)
-    } else if name.contains("skeleton")
-        || name.contains("king")
-        || name.contains("lord")
-        || name.contains("queen")
-        || name.contains("chief")
-        || name.contains("warlord")
-        || name.contains("commander")
-        || name.contains("warden")
-        || name.contains("admiral")
-        || name.contains("knight")
-        || name.contains("guardian")
-        || name.contains("matriarch")
-        || name.contains("sentinel")
-    {
-        Some(SpriteArchetype::Humanoid)
-    } else {
-        None
     }
 }
 
@@ -971,7 +340,6 @@ pub fn get_archetype_for_enemy(enemy_name: &str, zone_id: u32) -> SpriteArchetyp
     let clean_suffix = clean_name.split_whitespace().last().unwrap_or(clean_name);
 
     // 1. Try zone-based suffix matching
-    // If the suffix matched a known zone enemy, use that archetype
     if suffix_is_known_for_zone(zone_id, suffix) {
         return archetype_for_suffix(zone_id, suffix);
     }
@@ -989,174 +357,6 @@ pub fn get_archetype_for_enemy(enemy_name: &str, zone_id: u32) -> SpriteArchetyp
 
     // 4. Fall back to zone default
     zone_default_archetype(zone_id)
-}
-
-/// Checks if a suffix is a known zone enemy suffix.
-fn suffix_is_known_for_zone(zone_id: u32, suffix: &str) -> bool {
-    let s = suffix.to_lowercase();
-    match zone_id {
-        1 => matches!(
-            s.as_str(),
-            "beetle"
-                | "rabbit"
-                | "wasp"
-                | "boar"
-                | "serpent"
-                | "grub"
-                | "hare"
-                | "toad"
-                | "mantis"
-                | "sprout"
-        ),
-        2 => matches!(
-            s.as_str(),
-            "wolf"
-                | "spider"
-                | "bat"
-                | "treant"
-                | "wisp"
-                | "lynx"
-                | "moth"
-                | "hollow"
-                | "shambler"
-                | "wretch"
-        ),
-        3 => matches!(
-            s.as_str(),
-            "goat"
-                | "eagle"
-                | "golem"
-                | "yeti"
-                | "harpy"
-                | "ram"
-                | "condor"
-                | "troll"
-                | "bandit"
-                | "gargoyle"
-        ),
-        4 => matches!(
-            s.as_str(),
-            "skeleton"
-                | "mummy"
-                | "spirit"
-                | "gargoyle"
-                | "specter"
-                | "revenant"
-                | "shade"
-                | "lich"
-                | "cultist"
-                | "apparition"
-        ),
-        5 => matches!(
-            s.as_str(),
-            "salamander"
-                | "phoenix"
-                | "imp"
-                | "drake"
-                | "elemental"
-                | "cinderwyrm"
-                | "ashborn"
-                | "magmite"
-                | "hellhound"
-                | "infernal"
-        ),
-        6 => matches!(
-            s.as_str(),
-            "mammoth"
-                | "wendigo"
-                | "wraith"
-                | "bear"
-                | "wyrm"
-                | "moose"
-                | "banshee"
-                | "imp"
-                | "glacial"
-                | "revenant"
-        ),
-        7 => matches!(
-            s.as_str(),
-            "construct"
-                | "guardian"
-                | "sprite"
-                | "watcher"
-                | "golem"
-                | "shard"
-                | "crawler"
-                | "prism"
-                | "sentinel"
-                | "echo"
-        ),
-        8 => matches!(
-            s.as_str(),
-            "kraken"
-                | "shark"
-                | "naga"
-                | "leviathan"
-                | "siren"
-                | "eel"
-                | "ray"
-                | "lurker"
-                | "drowned"
-                | "hydra"
-        ),
-        9 => matches!(
-            s.as_str(),
-            "griffin"
-                | "djinn"
-                | "sylph"
-                | "roc"
-                | "wyvern"
-                | "zephyr"
-                | "pegasus"
-                | "manticore"
-                | "cloudwalker"
-                | "stormhawk"
-        ),
-        10 => matches!(
-            s.as_str(),
-            "titan"
-                | "colossus"
-                | "lord"
-                | "king"
-                | "champion"
-                | "warlord"
-                | "juggernaut"
-                | "thunderborn"
-                | "stormknight"
-                | "breaker"
-        ),
-        11 => matches!(
-            s.as_str(),
-            "beast"
-                | "horror"
-                | "fiend"
-                | "terror"
-                | "monster"
-                | "abomination"
-                | "void"
-                | "rift"
-                | "amalgam"
-                | "remnant"
-        ),
-        _ => false,
-    }
-}
-
-/// Checks if a suffix is from the generic dungeon enemy name pool.
-fn is_generic_suffix(suffix: &str) -> bool {
-    matches!(
-        suffix.to_lowercase().as_str(),
-        "orc"
-            | "troll"
-            | "drake"
-            | "crusher"
-            | "render"
-            | "maw"
-            | "beast"
-            | "fiend"
-            | "horror"
-            | "terror"
-    )
 }
 
 #[cfg(test)]
